删除之前车次之前增加了判断：判断删除的ids是否为空

+++ business/src/main/java/com/lxl/business/service/impl/DailyTrainServiceImpl.java
@@ -159,7 +159,9 @@ public class DailyTrainServiceImpl implements DailyTrainService {
         //先删除目标日期所有的车次,以及该车次的详情信息
         //先找出目标日期的车次的 id
         List<Long> dailyTrainIds = dailyTrainMapper.selectAimDateTrainIds(date);
-        dailyTrainMapper.deleteBatchIds(dailyTrainIds);
+        if (CollUtil.isNotEmpty(dailyTrainIds)) {
+            dailyTrainMapper.deleteBatchIds(dailyTrainIds);
+        }
         //删除目标日期存在的车站
         dailyTrainIds.forEach(dailyTrainId -> {
             Map<String, Object> columnMap = Map.of("daily_train_id", dailyTrainId);
